feat: Add common layout elements and structure
This adds the "block" element which is the half faded bordered element that is seen in several places in the design. With that I've also added the basic structure of the build detail pages. It's very bare bones, but I've tried to set up the HTML in a way that makes sense and that will hopefully stay once we get actual data in.

I've added some basic types, these are total placeholders, I expect these to change completely. It was just so there was a base type I could use for now.