feat: Add static statistics bars
This adds components and styling for statistic bars that are displayed on the detail pages. I don't yet know what the actual data will look like, or how we will calculate these values, but at least the styling is here!

By default the stats show the 5 main stats, but they can be expanded to show all stats that are available in the stats screen ingame.