feat: Builds list
This adds the builds list with individual build items. I'm not yet entirely satisfied with the view on mobile, and there's some bad breakpoints, but I will improve that in a separate PR later.

The builds are clickable but are not themselves <a> tags, because we can't have <a> tags inside of other <a> tags.

The content is all of course placeholder stuff.