feat: Add styling for armor and health to statistic bars
This is a somewhat in-between solution that adds armor and health values to statistics bars, without going to full segmented route like in-game. I do still want to add that at some point since they are easier to read, but for now this is a simple solution using a linear-gradient that still allows us to display the armor and health values in in some.