feat: Add round info to build detail page
This adds a static version of the round info to the detail page. The contains the individual rounds as well as the description. With a variant for mobile as well.

Once again I made some assuptions for types and suchs. Feel free to totally ignore these, these were just temporary to make development life a little easier.