feat: Add compounded build to sidebar on build detail page
This adds the "final" build to the sidebar of the detail page. This final build is derived from the items and powers per section, adding them up.
There's not yet any logic for sold items.
Types in the new utils/build function are not quite right because we don't have the actual types yet.