feat: Allow popover to be closed via outside click or escape
On desktop the interaction is a simple hover, and closing the popover doesn't require extra actions. With touch screens, however, that is a bit different. The popovers are opened on tap, and closed by tapping the same element.
This PR adds another way, and that is by tapping anywhere but the element itself. Or pressing escape, for good measure.