feat: Add item purchase order to build detail page
What you can buy per round is dependent on how much money you earned that round, as such the boxes per round might not always be entirely correct. To give a more simple overview of what items to buy in what order, this PR adds a rail at the top of the detail page that shows just that; items in purchase order. This is entirely derived from the items per round.