chore: Add page titles
Just some simple page titles. Not sure if these are final, but they look at little nicer than the default nothing.

I wasn't sure whether to go with "OW Armory" or "Overwatch Armory". In the end I went with "OW Armory" because it was shorter and prevents having "Overwatch" twice in the title on the homepage.