feat: Add load more button to latest builds
This lays the foundations for fetching of builds, along with fetching more builds with the press of a load more button.

Builds are fetched on the server and the client via +page.ts, which fetches from /api/builds/latest. This is passed down to +page.svelte and used for the initial value for the builds list. From here the user can press a load more button, which loads more titles fetched from the same api endpoint.

When a user navigates to a build and goes back using browsing history controls, the current state would be reset and the list would only contain the initial results again. This is where Svelte Snapshots come in. These allow you to store data when a user navigates away from a page, and restore it again when they return. Here we store the latestBuilds state, and the scrollPositon.