fix(vite): ensure only type data is imported from Prisma generated code
Prisma's generated client includes runtime code which relies on Node.js
internals. If the Prisma client is included in any client-side code
which runs on the browser, it will cause the entire page to fail to
render.

As a result, it is necessary to only use the prisma generated code for
client-side type checks, which the Vite build process will strip from
the final build. Note that using the client for the purposes of
server-side code should be fine.