feat: Add item purchase order to build detail page (#18)
## Description

What you can buy per round is dependent on how much money you earned
that round, as such the boxes per round might not always be entirely
correct. To give a more simple overview of what items to buy in what
order, this PR adds a rail at the top of the detail page that shows just
that; items in purchase order. This is entirely derived from the items
per round. Powers are not included in this, since they are not part of a
purchase order.

## Screenshots


![image](https://github.com/user-attachments/assets/05d3979c-c0dc-4133-a4c1-9c9131a8422b)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>