feat: Set up simple state for currentUser (#19)
## Description

We'll want to keep track of the logged in user somehow, and it needs to
work SSR and CSR. We can't store it in a store or a state rune, as these
are shared on the server, meaning data could leak from one user to the
next. Instead we store an auth token in a cookie. This cookie is used to
fetch the user from database. From here the user object is passed via
the layout and down to any other load function that might use it. This
same object is set via the context api via `setContext`. To make sure
this is updated whenever the state on the server changes (eg, the user
logs out or in), the main layout is wrapped in a `#key`. Whenever the
actual `data.currentUser` object is changed, the whole layout is
re-rendered, forcing `setContext` to be run again. This can then be used
by any component with `getContext`.

Just like that we have a global user state that works SSR and CSR and
isn't shared with the server.

The login and logout are behind an `/api` endpoint. This is mostly to
make it easier and more consistent to make calls to the server without
directly making calls to the database, the server will do that for us.
We make sure the user is updated on the server after calling the
endpoints by calling `invalidateAll` after logging in or out.

Of course the actual DB calls aren't included in this PR, so it's all
placeholder and made with some assumptions, but hopefully it's a good
foundation.

## Screenshots

This doesn't show much, but it shows that the state actually changes
when logging in and out. The actual endpoints are called and it isn't
just static.


![user-state](https://github.com/user-attachments/assets/d5e07268-c5de-4457-b443-d23731b96f09)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>