chore: Add page titles (#21)
## Description

Just some simple page titles. Not sure if these are final, but they look
at little nicer than the default nothing.

I wasn't sure whether to go with "OW Armory" or "Overwatch Armory". In
the end I went with "OW Armory" because it was shorter and prevents
having "Overwatch" twice in the title on the homepage.

Home | Detail
--- | ---

![image](https://github.com/user-attachments/assets/c1ad4691-bfc5-4269-9e4d-01a41a80a26b)
|
![image](https://github.com/user-attachments/assets/13eb2e85-1ab5-4e85-b6ef-de802b8ecdea)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>