feat: Account page (#22)
## Description

This merges into #19, which should be merged first.

This adds a simple account page which shows the user's builds and a log
out button. The page doesn't contain much else.
With this I've created generic button styling with a variant for the
primary and secondary button. I've removed some styling of other buttons
that now inherits from that generic styling too.

## Screenshots

Desktop | Mobile
--- | ---

![image](https://github.com/user-attachments/assets/11d02adf-4096-4f6e-beb8-c580a66289c6)
|
![image](https://github.com/user-attachments/assets/f62a981a-0b1f-4bdb-8a1c-ddf46b8ab1b0)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>