feat: App separate page for latest builds (#25)
## Description

This merges into #24, which should be merged first

This is a continuation of #24. In that PR we load more pages with a load
more button, but this isn't very SEO friendly, as all those additional
pages won't be picked up by search engines. This changes that by
changing the load more button to a link, rather than a button, which
links to `/latest` with the current page given. On that page we display
builds with a "next page" button. This page isn't paginated in much the
same way, as it's assumed this page is only used by search engines or
people with Javascript disabled.

## Screenshots


![latest-page](https://github.com/user-attachments/assets/f7a33e7d-459a-4ade-a1c7-e9b3418eccd3)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>