feat: Add card display for powers and items
Card displays show all detail for an item or power in one go, without requiring a popover hover. These are styled to look similar to how they look in-game.
This was mostly intended for the build form power select, but with us only having one section per round at the moment, it makes a lot of sense to show the full description. Especially while no one recognizes the icons yet. As such, I've applied the card to the build detail pages as well.