feat: Build form (#29)
## Description

This merges into #25, which should be merged first

This PR adds the form used to create and edit builds. It's quite a large
PR unfortunately, because I kept running into things I didn't account
for. Let's start from the top.

- When creating a new build, a user can select a hero. When editing a
post, the hero is locked.
- A user can input a title
- A user can input a short description, which serves as the introduction
- A user can select items and powers per round
- Each round presents a grid of items and powers similar to the view
in-game.
- If an item is picked in round 1, it can be sold in any subsequent
round
- When this is done, the item is added to that round. There is nothing
that indicates that that item is sold, except for the number of times an
item occurs. If an item occurs twice, that means it was bought the first
time, sold the second time. Once more, and the item is bought again.
- A user can go back to and edit any previous round. If an item or power
is selected in round 5, and the user goes back to round 3 and selects
the same items and powers, they are removed from round 5 (or any round
after 3 that they appear in)
- If a user changes hero all items unique to the hero are removed from
each round, and all powers are removed altogether
  - Powers can only be selected on round 1, 3, 5 and 7
- A user can input some short notes per round
- After the rounds display a concise order of purchased powers and items
is shown, with selling order also visible
- A user can input a larger description explaining the overall build
- A user can save their build
- When a new build is made it is posted with a POST request to
`/api/build/post`
- When a build is edited the form is post with a PATCH request to
`/api/build/patch`
- Each request has a number of validations that are done on the server.
Thing title length, hero selection, number of rounds given, etc.
- If validations fail an error is returned, the user is scrolled back
up, and the error is shown. Currently only 1 error can show at a time,
even though a form might contain multiple.
- If validations succeed, nothing really happens because this isn't
connected to the database in any way just yet.
  
What this doesn't contain yet:
- Items are placeholders, there's no filtering for items or powers for
specific heroes, or even for the different "Weapon, ability, survival"
tabs.
- The powers tab is not final. I am adding a full card view for both
items and powers, that isn't included with this PR just yet.
- The form is entirely static, it isn't being posted anywhere real
- The textarea would do well with some sort of markdown editor,
especially the large text area.

## Screenshots

New | Edit
--- | ---

![image](https://github.com/user-attachments/assets/f2fc7be5-d309-4d49-a194-426520368f3e)
|
![image](https://github.com/user-attachments/assets/60c21987-0bcf-4c4e-9ed2-648d108c9e75)


![new-build](https://github.com/user-attachments/assets/015ebf73-8c98-41da-a5e1-488756244745)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>