feat: Add card display for powers and items (#30)
## Description

This merges into #29 

Card displays show all detail for an item or power in one go, without
requiring a popover hover. These are styled to look similar to how they
look in-game. This was mostly intended for the build form power select,
but with us only having one section per round at the moment, it makes a
lot of sense to show the full description. Especially while no one
recognizes the icons yet. As such, I've applied the card to the build
detail pages as well.

## Screenshots

### Desktop

![image](https://github.com/user-attachments/assets/6b5fd3d4-5762-4c58-9233-3e6ea56f9c66)

![image](https://github.com/user-attachments/assets/7440d95b-ac62-488b-8da4-70dcc1ac412e)

### Mobile

![image](https://github.com/user-attachments/assets/9907de8a-255c-40ca-aefa-0e5281bdde16)

### Form
Unselected | Previous round shown + selected
--- | ---

![image](https://github.com/user-attachments/assets/8fe0bdc1-f5f7-49f6-8476-8432fe5d5b7f)
|
![image](https://github.com/user-attachments/assets/9aea74e7-c15d-478a-9c19-7d1eb0d935f3)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>