feat: Account for sold items in compounded build (#32)
## Description

This merges into #29 

The compounded build would show all items, including items that may have
been sold. This PR cares care of that in both the shown items and the
total build cost.

## Screenshots

This is probably a bit confusing to look at, but I buy 2 items in round
1, sell them both for round 2, and buy 3 items (one of which I also had
in round 1) in round 3.


![compounded-build-sell](https://github.com/user-attachments/assets/e76c1375-a204-4a03-bf7c-9021bfc5c14d)

![compounded-build-sell-list](https://github.com/user-attachments/assets/de50c17a-4cc5-484a-93cd-92a500d4233a)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>