feat: Auth (#36)
## Description

This PR adds authentication via Battle.net OAuth. This is all handled
through auth.js with an adapter for Prisma. This required a pretty
specific schema for user models, most fields we aren't using at all, but
I still had to include them. The good this here is that it sets us up
for adding authentication via other providers or via email.

The sign in buttons work via a form with SvelteKit actions, this is all
done through auth.js and I didn't have to set anything of this up
myself.

The currentUser object is still passed in the same way. The session is
passed via the main layout file and used via the context store. This
works server side. This session is saved for 30 days. I'm not yet sure
how to extend this.

## Screenshots

![oauth](https://github.com/user-attachments/assets/003b0087-d99b-4a34-97aa-987a3805bfd9)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>