feat: Search (#37)
## Description

This PR adds search via full-text search from Prisma. Search results are
search for on either the hero name or the build title. The query is
inclusive(?) in that results require all words to be present. If you
search "Cool build", both "Cool" and "build" need to be present in the
title.

The search form itself is placed in the main navigation, but it is
unstyled. The styles it would rely on are in #29, and I didn't want to
extend that chain any more.
The form is not showing results as you type, instead only showing
results when you press enter. This is done through a simple navigate,
where the query param is used in the server endpoints.

## Screenshots


![search](https://github.com/user-attachments/assets/e769fb07-6314-449e-9b20-add63d1bf750)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>