feat: Hero and user pages (#38)
## Description

This merges into #36 

This PR adds pages for all builds for a particular hero and all builds
for a particular user. With this I also refactored some other bits to be
a bit more re-usable, readable, or users to use. There still is quite a
lot of repetition between pages, and that's largely because of Snapshots
which have to be right inside a +page file and can't be inside a shared
component. Bit annoying, but something to worry about some other day.

There's a 404 for both user and hero pages. One if the user doesn't
exist, the other if the hero name doesn't exist.

## Screenshots

### Heroes

![heroes](https://github.com/user-attachments/assets/268849fa-6d04-406a-a3c4-2c6816cf5f9f)

### User page

![image](https://github.com/user-attachments/assets/452e9d5d-16bf-4516-a929-a6bda1ed1266)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>