feat: Use id-slug routes for build pages (#54)
## Description

We want to use build `id`s to fetch data and to create unique data, but we want to use build titles for create SEO and human friendly URLs. This PR create the best of both worlds by moving the build detail path to `[id]-[slug]`. The id is used to fetch data, and the slug to used purely decoratively. When you enter a page the title of the build is sluggified, this is compared to the actual slug in the URL, and if they don't match the user is redirected. This ensures that the path can't be altered by the user (and thus creating SEO duplicates, which is a no-no in SEO land), and it will automatically redirect after the user has changed the title of their build.

## Screenshots

This is me showing off the redirect

![build-path-redirect](https://github.com/user-attachments/assets/e2b80e60-5d0c-45ed-b65d-6c3e20dd8aae)