feat: Turbolinks-like navigating indicator
When waiting for a new page to load, it can seem like nothing is happening. Of course usually it's fast enough to where it's not a problem, but sometimes it's not.
This PR adds a loading indicator similar to that of Turbolinks to make it obvious that _something_ is happening.