feat: Randomize background images
For a little bit of visual interest, this change adds random background images to each page. So ensure that the background is the same between SSR and CSR, I've used `seedrandom` to always generate the same background image per path.

Is this a cool feature, or is it just distracting and weird?