fix: Additionally sort items by name
All general items are sorted by cost and then by name. For hero specific items that's less so the case, _most_ items seem to be sorted by name, but not all. So even this sorting isn't quite right, but it is one step closer. The only proper sorting seems to be to do it manually, which I'm not sure is worth the extra effort. Largely because hero specific items aren't exactly plentyful anyway.