style: Mobile navigation
The currently navigation did not fit on mobile. I still wanted to keep it simple, so I shrank down everything a bit, hiding things that are too large, replacing them with icons. Not the best solution, and it would be nicer to move things to a proper fold out menu, but that's for another time.