feat: Add item stat mods to description (#64)
## Description

Fixes #65 

This PR adds item stats to descriptions. I couldn't figure out how to do
the full type for items to include StatMod and Stat, so a little help
there would be much appreciated!

## Screenshots

_ | _
--- | ---

![image](https://github.com/user-attachments/assets/34da8ced-8617-413d-bad8-9c30f15874b1)
|
![image](https://github.com/user-attachments/assets/d8165915-2571-40a1-b4cb-ff0f3486ee6f)

![image](https://github.com/user-attachments/assets/e9975266-5822-4b34-8139-9d072a9e1ea0)
|
![image](https://github.com/user-attachments/assets/a35768a7-42a9-4643-a519-ba63b55ad9a7)

![image](https://github.com/user-attachments/assets/cf5500b3-ad0d-444e-ba65-ff1de991bf03)
|
![image](https://github.com/user-attachments/assets/284402fe-5d65-4ca2-b432-999ff4a1399b)

![image](https://github.com/user-attachments/assets/9cf44b2a-d434-4547-9c4a-b1e5edfceba5)

This one is completely empty for some reason? It doesn't have any stat
mods in the database. It's supposed to be +25% starting ultimate charge,
but we don't seem to have that stat.


![image](https://github.com/user-attachments/assets/a75a462d-f5ad-4d71-8347-489a040da635)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>