feat: Mark numbers in descriptions for readability  (#66)
## Description

This merges into #64 

To match the in-game formatting, numbers in the description of items are
wrapped with `<mark>`, turning them white. It's possible this will match
some numbers that are not supposed to be marked, but I haven't found any
yet. Perhaps if Soldier: 76 ever gets mentioned in a description, we'll
have a revisit how this is done. For now simple and stupid seems just
fine.

## Screenshots


![image](https://github.com/user-attachments/assets/85531998-e699-4cb3-9190-657f0315ca31)

![image](https://github.com/user-attachments/assets/1b75a644-a428-4853-9ea9-8d03d004c1f6)

![image](https://github.com/user-attachments/assets/c475a328-bcd5-422a-879b-27e694f39234)

![image](https://github.com/user-attachments/assets/b85df6ca-700a-4ace-ac33-e3f4e59e2f01)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>