feat: Randomize background images (#73)
## Description

For a little bit of visual interest, this change adds random background
images to each page. So ensure that the background is the same between
SSR and CSR, I've used `seedrandom` to always generate the same
background image per path.

Is this a cool feature, or is it just distracting and weird?

## Screenshots


https://github.com/user-attachments/assets/fed0c062-9dfc-4eb9-a2c2-afe5a21054a1