feat: Derive stats from items (#77)
## Description

This derives all percentage stats from the actual build items. Health
isn't included yet, as that's quite separate and we're missing a whole
bunch of stats to make this work. We seem to be lacking shields and
armor.

The stats array is a bit verbose, but it helps us keep the correct
order.

## Screenshots

![stats-change](https://github.com/user-attachments/assets/620ffad2-cd4f-4e3c-8382-1dc84effb9d7)