feat: Add full items and powers pages (#78)
To help us debug and find incorrect powers/items, this PR adds a powers
and items page that simply lists all powers or items, grouped by hero.

This isn't really meant to be a public page, so it's super basic and not
all that well made. Maybe one day these could be public facing, with a
little search bar too, that'd be neat.

## Screenshots

Powers | Items
--- | ---

![image](https://github.com/user-attachments/assets/6f387275-3c38-4ae6-860c-f4428ee5826f)
|
![image](https://github.com/user-attachments/assets/d0810ae0-c34e-4779-a95b-0255da406b19)