style: Mobile navigation (#82)
## Description

The currently navigation did not fit on mobile. I still wanted to keep it simple, so I shrank down everything a bit, hiding things that are too large, replacing them with icons. Not the best solution, and it would be nicer to move things to a proper fold out menu, but that's for another time.

## Screenshots

This is a pretty small size, smaller than most modern phones.

Signed out | Signed in
--- | ---
![image](https://github.com/user-attachments/assets/8d8c636a-6998-433c-9b72-298b5d3f4496) | ![image](https://github.com/user-attachments/assets/9efb513e-3787-4c6a-9f60-968d953a247c)