fix: Filter builds by their IDs in lists (#84)
## Description

Builds in lists need to be unique, otherwise the #each loops run into
errors. In theory builds are always unique, but there are some scenarios
where they wouldn't be. This includes builds being added while a user is
on one page, and navigating to a new page would now include one of the
builds of the previous page because everything shifted one down. This is
fixed by filtered all builds in a list by their ID, only returning
unique builds.

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>