feat: Indicate progression for navigating indicator (#97)
## Description

When a page takes a while to load, the navigating indicator hits 20% and
then stays there, making it feel like nothing is happening. This change
will still slowly progress the bar, filling up to 80% over 20 seconds.
It's totally fake of course, but it's something to show that something
is happening. This once again is just like how Turbolinks does it.

## Screenshots


![navigating-indicator-progress](https://github.com/user-attachments/assets/e4621139-b4e1-4108-85df-bc6fb577ee18)

Co-authored-by: CactusPuppy <puppycactus@gmail.com>