feat: Add "similar to" and heroes to bottom of build (#98)
## Description

This adds a similar to and heroes section after all the build details.
The similar to section is fetched async on mount, not on the server. All
this does for now is fetch builds for the same hero, making it not
terribly similar to, but while we have little to no content, it will do.
In the future we can grab builds with similar tags (#94).
For this to work I had to make some changes. Builds had to be derived
from `data` as otherwise navigating from build to build would not
properly update the state. Previously the build was passed via the
context api, but this was unnecessary as any data from a layout file is
also available in page files.

## Screenshots


![image](https://github.com/user-attachments/assets/aab4099d-5188-4176-8a5d-150d0669f556)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>