feat: Items powers search (#99)
## Description

This merges into #78 

In #78 I added pages for items and powers with the only goal for us to
have a nice overview. I added a search field to this page, and then
realized that this page is actually really quite useful, so we might as
well expose it to users!

In doing so I cleaned up some stuff. Previously the two pages were
essentially copy pages from each other. Here I cleaned that up by moving
the grouping and filtered to a single function that is passed via
context.

Searching happens by searching on the full object, stringified, which is
copied from what I did here for the same reason as explained:
https://github.com/CactusPuppy/owarmory/pull/80

Rendering of the items works by re-rendering the entire grid whenever
the query changes. Super inefficient and stupid, and it can feel a
little slow at times, but I don't super care at the moment!

The navigations links on the homepage don't look great. I'm trying to
think of a more pleasing design, but for now it's function over form.

## Screenshots


![talents-search](https://github.com/user-attachments/assets/5a106e47-cd44-4003-9e70-01e05be7d9fe)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>