feat: Talent text previews (#101)
## Description

Merges into #100 

This one was a bit hard to come up with a proper solution for. I ended
up creating a separate component that parses the text it's given,
rendering regular text as text and given {{tags}} as Svelte components.
This allows us to insert items as {{Item Name}}. This item name is in no
way verified, and thus it can be invalid. When a user hovers over a
talent preview the talent is fetched. Currently it's not too smart. It
fetches an item first, and if that fails it fetches a power second. If
that fails a message is shown that no talent or item was found.

## Screenshots


![talent-previews](https://github.com/user-attachments/assets/a610f435-56ee-4334-b931-6d546ccb58a9)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>