feat: Add markdown preview to description input (#103)
## Description

Merges into #101 

To make it easier to see what your markdown is doing, we now show a
preview next to the form input. It updates live as you type which is
done through the use of a `#key` block, forcing that element to
re-render every time the value is updated.

## Screenshots


![markdown-preview](https://github.com/user-attachments/assets/d5cd7830-ae48-4d5a-a892-a17225a6350e)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>