fix: Flip popover to top when it doesn't fit on screen
When the popover doesn't fit on the page, it break out of the page but you can never actually scroll to it. This fixes that by opening the popover to the top instead. This is dependant on screen position, not the position on the page. In other words, it always tries to fit it on your screen, not just on the page.