fix: Don't include full statmods in stats
For a reason that I'm not 100% sure on, the `stat` field in `statMods` included `statMods` itself, which then seemed to include all `statMods` relations ever created for that stat? This resulted in builds having giant arrays of irrelevant statmods on each stat that only grows as more builds are created. JSON returns for builds could easily be be up to 1MB already, and overview pages could be 5MB+ already.

Anyway, this fixes that by no longer including `statMods` within `stat`, only returning the relevant fields instead.