feat: Add compounded build to build form (#159)
## Description

To make it easier to see where you're at with your build, this adds the
CompoundedBuild component from the actual build pages to the build form.
The layout gets a bit cramped like this, but it adds a fair bit of
clarity. Items can be sold via the compounded build. Powers can not,
which matches ingame behaviour.

## Screenshots


![compounded-build-form](https://github.com/user-attachments/assets/000477c0-a737-428e-87aa-d9f7df1f535e)

Desktop | Mobile
--- | ---

![image](https://github.com/user-attachments/assets/65d4689b-49e5-4f1d-b3d0-f5fd57729595)
|
![image](https://github.com/user-attachments/assets/ed85a078-bd23-4c4c-a9a5-9bab6c86661c)

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>