Add way to reseed data with updated stadium while preserving existing data + add Freja data (#157)
Updates the seed script so that it can be re-ran with updated Stadium
data to update all the data. This should be properly tested on a staging
database that has live data on it.


**Overview of changes**
* Adds migration to update DB schema to add `gameGuid` to Power, Item,
and StatMod tables. As well as updating all existing Items/Powers with a
guid.
* Updates seed script to update existing items/powers or create new
ones. Updating is done based on game guid.
* When updating items, it also updates the stats and stat mods for the
items.
- New stats are created if they don't already exist based on the stat
name (same as before)
  - Stat Mods are updated based on game guid otherwise created
- Any existing stat mods that weren't updated during the run get deleted
(see note)
* Refactors how custom data works, most custom data is no longer needed
as the updated stadium data includes the previously missing data.
However, there is now a basic system that allows you to manually add
custom data to a talent for cases where custom data is needed. It's very
basic though and only really there to handle the 2 unique cases where
additional custom buffs are required.
 * Updates Stadium Data with latest changes from the game
* Adds Freja to the heroes data and includes their stadium data and
images

### Important Note about Stat Mods
The first time running the seed script to update data will **recreate
all StatMods** and **delete all the existing ones**. In order for it to
properly update StatMods, it needs the game guid to be stored which
won't exist on a pre-seeded database. But once the mods have the guids,
they will be updated instead of remade.

There is one small exception here: only stat mods with a game guid will
be updated, any custom stats added will get recreated and the previous
ones deleted everytime the script is ran. This is because they have no
game guid so no easy way to track where they came from. Currently, there
is only 4 custom stats so it's not really an issue. Technically though
you could just give them a fake id (maybe)

This isn't really an issue anyway because nothing inherently depends on
the stat mod ids so the ids changing shouldn't affect anything.

### How to run/test
On a pre-seeded database, you should just be able to run these commands:
```bash
# apply latest migration
pnpm prisma migrate dev
# run seed script again to update and add new data
pnpm prisma db seed
```

---------

Co-authored-by: CactusPuppy <puppycactus@gmail.com>