feat: show time since build update in list (#169)
This PR adds text which informs the user about when a given build was
last updated in the build list, along with changing "Last updated" text
to use relative units instead of absolute dates. Absolute dates are
still visible on hover.

**Motivation**:
As Stadium is frequently balanced, builds which may have been top-tier
in previous patches may become less powerful (intentionally so!) as the
team balances the game. Adding information about when a build was last
updated enables players to make more informed decisions about which
builds to examine further.

### Example list

![image](https://github.com/user-attachments/assets/6b3eec95-c3ec-4b81-b6f9-9c11390a3ec5)

### Updated build view page header

![image](https://github.com/user-attachments/assets/efbd9b43-d344-4464-8547-23a33b55b2b6)